requirements.txt
pandas==1.3.5
numpy
streamlit==1.11.0
pykalman==0.9.5
pillow==9.4.0
scipy==1.7.3
matplotlib==3.5.3
flake8==6.0.0